+++ cg/cli/workflow/fastq/base.py
@@ -6,7 +6,7 @@ from cg.cli.workflow.commands import ARGUMENT_CASE_ID
 from cg.cli.workflow.fastq.fastq_service import FastqService
 from cg.constants.constants import DRY_RUN, Workflow
 from cg.meta.workflow.analysis import AnalysisAPI
-from cg.services.quality_controller import QualityControllerService
+from cg.services.sequencing_qc_service import QualityControllerService
 from cg.store.store import Store
 
 LOG = logging.getLogger(__name__)
