+++ cg/meta/workflow/analysis.py
@@ -28,7 +28,7 @@ from cg.meta.workflow.fastq import FastqHandler
 from cg.models.analysis import AnalysisModel
 from cg.models.cg_config import CGConfig
 from cg.models.fastq import FastqFileMeta
-from cg.services.quality_controller import QualityControllerService
+from cg.services.sequencing_qc_service import QualityControllerService
 from cg.store.models import Analysis, BedVersion, Case, CaseSample, Sample
 
 LOG = logging.getLogger(__name__)
