--- cg/services/quality_controller/__init__.py
@@ -1,1 +1,0 @@
-from cg.services.quality_controller.quality_controller_service import QualityControllerService
