+++ cg/services/sequencing_qc_service/__init__.py
@@ -1,0 +1,1 @@
+from cg.services.sequencing_qc_service.sequencing_qc_service import QualityControllerService
