+++ cg/services/sequencing_qc_service/sequencing_qc_service.py
@@ -1,7 +1,7 @@
 from typing import Callable
 
 from cg.store.models import Case, Sample
-from cg.services.quality_controller.quality_checks.checks import (
+from cg.services.sequencing_qc_service.quality_checks.checks import (
     run_quality_checks,
     get_sequencing_quality_check_for_case,
     get_sample_sequencing_quality_check,
